Move layout components into own folder
