+++ src/App.vue
@@ -1,7 +1,7 @@
 <script setup>
   import { provide } from 'vue';
-  import LayoutFooter from './components/LayoutFooter.vue';
-  import LayoutHeader from './components/LayoutHeader.vue';
+  import LayoutFooter from './components/layout/LayoutFooter.vue';
+  import LayoutHeader from './components/layout/LayoutHeader.vue';
   import { layout } from './state';
 
   provide('layout', layout);
